fix : entity id 데이터타입을 int => Long으로 변경

+++ src/main/java/com/zerobase/morse/entity/ChatContent.java
@@ -9,7 +9,7 @@ public class ChatContent {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int chatNo;
+    private Long chatNo;
     private int roomId;
     private String email;
     private String content;
